# golang/internal/server/api/openapi.yaml
openapi: 3.0.3
info:
  title: wow-converter
  version: "1.1.0"
  description: |
    This server reads a World of Warcraft installation and hands you the files.
    Model, icon, and map calls convert those files into the formats Warcraft III uses.
    Sound calls do not convert anything. They return the original audio.
    GET /api/sound/transcripts is a separate list of spoken lines for those files.

    There is no login. Anyone who can open the port can call every endpoint.
    A server you run yourself listens on 127.0.0.1 port 3001 unless HOST or PORT was changed.
    A public shared server listens on every interface and refuses calls that would
    change the game install, delete files, or build a map.

    ## How a call looks

    Send JSON as application/json. File paths inside the game use forward slashes.
    When a path is part of the URL, it is the remainder of that URL and it contains slashes.
    Percent-encoding those slashes is accepted.

    Lists are not paginated. Pass search to keep the body small. An empty search is the whole catalog.

    A 200 that is audio, PNG, or a zip is the file itself, not JSON. A failure on that same URL is JSON.
    Read the status code before deciding how to read the body. 304 means the body you already have is still current, and this response has no body.

    Opening an edition, and the first list after that, can take minutes. A client that gives up after a few seconds will fail a call that is still working.

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/sound:
    get:
      tags: [Sounds]
      summary: List original audio files
      description: |
        Use this to find a sound, then stream or zip it. The bytes are the game's own .mp3, .ogg, or .wav. Nothing is converted.
        Send search. An empty search returns every sound, which is large.
        Requires an open game edition.
      operationId: listSounds
      parameters:
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Build'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Matching sounds, sorted by path.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameFile'
        '304':
          $ref: '#/components/responses/NotModified'
        '500':
          $ref: '#/components/responses/Error'
  /api/sound/transcripts:
    get:
      tags: [Sounds]
      summary: Spoken lines for sound files
      description: |
        Call this once and keep the map. Keys are file data ids, as strings because JSON object keys are strings.
        The value is the spoken line. A file with no line is absent.
        This does not read the open game edition, and it ignores search and build.

        The server reads resources/transcripts.final.zip at startup. The zip holds transcripts.final.jsonl.
        Each line in that file is {"id": 554181, "text": "Bow to your master.", "ai": 1}.
        ai 0 means the words are the game's own BroadcastText. ai 1 means they were generated.
        When both exist, the file stores the game line. This response drops ai and returns only the text.

        A missing zip still returns 200 with an empty object.
        The response has an ETag. Send it as If-None-Match to get 304 when the map has not changed.
      operationId: listSoundTranscripts
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: File data id to spoken line.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: string }
              example:
                "554181": "Bow to your master."
                "558782": "BY FIRE BE PURGED!"
        '304':
          $ref: '#/components/responses/NotModified'
  /api/sound/{fileDataID}:
    get:
      tags: [Sounds]
      summary: Play or save one sound
      description: |
        Returns the original audio so a player can stream it. Range requests work, so seeking works.
        download=1 adds a save-as header that uses the file's own name.
        The id must come from the sound list. This route will not serve models or images.
      operationId: getSound
      parameters:
        - name: fileDataID
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
        - name: download
          in: query
          schema: { type: string, enum: ['1'] }
          description: Set to 1 to download instead of playing inline.
        - $ref: '#/components/parameters/Build'
      responses:
        '200':
          description: Audio bytes.
          content:
            audio/mpeg: { schema: { type: string, format: binary } }
            audio/ogg: { schema: { type: string, format: binary } }
            audio/wav: { schema: { type: string, format: binary } }
        '206':
          description: A byte range of the audio.
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
  /api/sound/zip:
    post:
      tags: [Sounds]
      summary: Download many sounds as one zip
      description: |
        Temporary: only 1 id is accepted so downloads stay one file at a time.
        Restore bulk zip by deleting the one-file check in sound.go and setting maxItems back to 1000.
        Use this when you want the files on disk, not a player. Send 1 id from the sound list.
        Paths inside sounds.zip match the paths from the list.
        A file that is not in the sound list, or whose bytes cannot be read, is left out of the zip.
      operationId: zipSounds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileDataIDs]
              properties:
                fileDataIDs:
                  type: array
                  minItems: 1
                  maxItems: 1
                  items: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Zip archive named sounds.zip.
          content:
            application/zip:
              schema: { type: string, format: binary }
        '400':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
...
  /api/download:
    post:
      tags: [Files]
      summary: Download exported files as one zip
      description: |
        Use this to take home the files a model or icon export wrote.
        files are the relative paths from that result. source browse reads the preview folder. Anything else, including omitting source, reads the kept-export folder.
        The zip is named after the first model in the list, or assets.zip when there is no model.
        A missing file, or a path that leaves the export folder, is 400.
      operationId: downloadAssets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  minItems: 1
                  items: { type: string }
                source:
                  type: string
                  enum: [export, browse]
                  description: Defaults to the normal export folder.
      responses:
        '200':
          description: Zip of the requested files.
          content:
            application/zip:
              schema: { type: string, format: binary }
        '400':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    Search:
      name: search
      in: query
      description: |
        Why: the full catalog is tens of megabytes. How: words separated by spaces.
        Every word must appear in the file path or in the decimal file id. Case is ignored.
        Leave it out to get the full catalog.
        A shared server returns 403 when search is set. File lists there are only available unfiltered.
      schema: { type: string }
    Build:
      name: build
      in: query
      description: |
        Why: a cached image from one game edition must not be reused after you open another.
        How: send buildKey from GET /api/get-config. When the edition changes, this value changes.
      schema: { type: string }
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: ETag from an earlier response of the same URL. If the body is unchanged the server returns 304.
      schema: { type: string }
...
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotModified:
      description: The cached copy is still current.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    GameFile:
      type: object
      required: [fileDataID, fileName]
      properties:
        fileDataID:
          type: integer
          description: The integer the game uses for this file. Pass it to the sound routes.
        fileName:
          type: string
          description: Path inside the game, with forward slashes. Pass it to texture preview and as a local model source.
...